# Frontend_LearningSwap/openapi_remote.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/register":{"post":{"summary":"Register a new user","operationId":"register_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"summary":"User Login","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages/{room_id}":{"get":{"summary":"Get Chat Room Messages","operationId":"get_messages_messages__room_id__get","parameters":[{"name":"room_id","in":"path","required":true,"schema":{"type":"integer","title":"Room Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai":{"post":{"summary":"Interact with the AI Assistant","operationId":"ai_chat_ai_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/onboarding/skills":{"post":{"summary":"Add User Skills Onboarding","operationId":"add_user_skills_onboarding_skills_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSkillsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/swipe":{"post":{"summary":"Record a User Swipe","operationId":"swipe_user_swipe_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InteractionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feed/{user_id}":{"get":{"summary":"Get User Feed","operationId":"get_feed_feed__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Feed Feed  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/matches/{user_id}":{"get":{"summary":"Get User Matches","operationId":"get_matches_matches__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Matches Matches  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"summary":"Get User Profile","operationId":"read_user_profile_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update User Profile","operationId":"edit_user_profile_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_edit_user_profile_users__user_id__put"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AIRequest":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"AIRequest"},"ActionEnum":{"type":"string","enum":["like","pass"],"title":"ActionEnum"},"Body_edit_user_profile_users__user_id__put":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"bio":{"type":"string","title":"Bio"},"phone":{"type":"string","title":"Phone"},"foto":{"type":"string","format":"binary","title":"Foto"}},"type":"object","title":"Body_edit_user_profile_users__user_id__put"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InteractionRequest":{"properties":{"user_from_id":{"type":"integer","title":"User From Id"},"user_to_id":{"type":"integer","title":"User To Id"},"action":{"$ref":"#/components/schemas/ActionEnum"}},"type":"object","required":["user_from_id","user_to_id","action"],"title":"InteractionRequest"},"UserCreate":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"phone":{"type":"string","title":"Phone"}},"type":"object","required":["first_name","last_name","email","password","phone"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserProfileResponse":{"properties":{"user_id":{"type":"integer","title":"User Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"avatar_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Url"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"skills_to_learn":{"items":{"type":"string"},"type":"array","title":"Skills To Learn","default":[]},"skills_to_teach":{"items":{"type":"string"},"type":"array","title":"Skills To Teach","default":[]}},"type":"object","required":["user_id","first_name","last_name","email"],"title":"UserProfileResponse"},"UserSkillsRequest":{"properties":{"user_id":{"type":"integer","title":"User Id"},"learn_skills":{"items":{"type":"string"},"type":"array","title":"Learn Skills"},"teach_skills":{"items":{"type":"string"},"type":"array","title":"Teach Skills"}},"type":"object","required":["user_id","learn_skills","teach_skills"],"title":"UserSkillsRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}